pyproject.toml
[tool.poetry]
name = "millie"
version = "0.1.0"
description = ""
authors = ["hyeseong-lee <hyeseong43@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.7"
PyJWT = "^2.0.1"
bcrypt = "^3.2.0"
requests = "^2.25.1"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
